refactor(level5): 5-1 패키지 구조 개선 추가수정
