+++ src/main/java/org/example/expert/domain/user/controller/UserAdminController.java
@@ -1,7 +1,7 @@
 package org.example.expert.domain.user.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.example.expert.domain.auth.annotation.AuthLogging;
+import org.example.expert.global.config.logging.annotation.AuthLogging;
 import org.example.expert.domain.user.dto.request.UserRoleChangeRequest;
 import org.example.expert.domain.user.service.UserAdminService;
 import org.springframework.web.bind.annotation.PatchMapping;
